legacy/requirements.txt
fastapi==0.112.0
Jinja2==3.1.4
loguru==0.7.2
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
slowapi==0.1.9
SQLAlchemy==2.0.32
starlette==0.37.2
uvicorn==0.30.5
itsdangerous==2.2.0